requirements.txt
fire==0.4.0
pandas==1.3.3
python-dotenv==0.19.0
python-bitvavo-api==1.2.2
xlsxwriter==3.0.1